+++ src/components/Pen.vue
@@ -19,13 +19,28 @@ const markmapStore = useMarkmap()
       class="absolute bottom-0 flex justify-center bg-gray-900/20 w-full z-50"
     >
       <div class="px-6 p-2 flex justify-wrap flex-wrap gap-0.5">
-        <button class="pen-button" id="undo" aria-label="Undo" title="Undo">
+        <button
+          class="pen-button"
+          id="undo"
+          aria-label="Undo"
+          title="Undo [mod+z]"
+        >
           ↩️
         </button>
-        <button class="pen-button" id="redo" aria-label="Redo" title="Redo">
+        <button
+          class="pen-button"
+          id="redo"
+          aria-label="Redo"
+          title="Redo [mod+shift+z]"
+        >
           ↪️
         </button>
-        <button class="pen-button" id="clear" aria-label="Clear" title="Clear">
+        <button
+          class="pen-button"
+          id="clear"
+          aria-label="Clear"
+          title="Clear [c]"
+        >
           🗑
         </button>
         <div class="mt-1 mx-2 opacity-25">/</div>
@@ -33,7 +48,7 @@ const markmapStore = useMarkmap()
           class="pen-button"
           id="m-arrow"
           aria-label="Arrow"
-          title="Arrow"
+          title="Arrow [a]"
         >
           ↗
         </button>
@@ -41,14 +56,24 @@ const markmapStore = useMarkmap()
           class="pen-button active"
           id="m-stylus"
           aria-label="Stylus"
-          title="Stylus"
+          title="Stylus [s]"
         >
           ✍️
         </button>
-        <button class="pen-button" id="m-draw" aria-label="Draw" title="Draw">
+        <button
+          class="pen-button"
+          id="m-draw"
+          aria-label="Draw"
+          title="Draw [d]"
+        >
           ✏️
         </button>
-        <button class="pen-button" id="m-line" aria-label="Line" title="Line">
+        <button
+          class="pen-button"
+          id="m-line"
+          aria-label="Line"
+          title="Line [f]"
+        >
           ⁄
         </button>
 
@@ -56,7 +81,7 @@ const markmapStore = useMarkmap()
           class="pen-button"
           id="m-rectangle"
           aria-label="Rect"
-          title="Rect"
+          title="Rect [r]"
         >
           ⃞
         </button>
@@ -64,7 +89,7 @@ const markmapStore = useMarkmap()
           class="pen-button"
           id="m-ellipse"
           aria-label="Ellipse"
-          title="Ellipse"
+          title="Ellipse [t]"
         >
           ⃝
         </button>
@@ -72,12 +97,18 @@ const markmapStore = useMarkmap()
           class="pen-button"
           id="m-eraseLine"
           aria-label="Eraser"
-          title="Eraser"
+          title="Eraser [e]"
         >
           🧹
         </button>
         <div class="mt-1 mx-2 opacity-25">/</div>
-        <button class="pen-button" data-color="#000000" id="pen-color-black">
+        <button
+          class="pen-button"
+          data-color="#000000"
+          id="pen-color-black"
+          aria-label="Black"
+          title="Black [mod+1]"
+        >
           ​⚫️​
         </button>
         <button
@@ -85,7 +116,7 @@ const markmapStore = useMarkmap()
           data-color="#ed153d"
           id="pen-color-red"
           aria-label="Red"
-          title="Red"
+          title="Red [mod+2]"
         >
           ​🔴​
         </button>
@@ -94,7 +125,7 @@ const markmapStore = useMarkmap()
           data-color="#ed9a26"
           id="pen-color-orange"
           aria-label="Orange"
-          title="Orange"
+          title="Orange [mod+3]"
         >
           ​🟠​​
         </button>
@@ -103,7 +134,7 @@ const markmapStore = useMarkmap()
           data-color="#ede215"
           id="pen-color-yellow"
           aria-label="Yellow"
-          title="Yellow"
+          title="Yellow [mod+4]"
         >
           ​​🟡​​
         </button>
@@ -112,7 +143,7 @@ const markmapStore = useMarkmap()
           data-color="#30bd20"
           id="pen-color-green"
           aria-label="Green"
-          title="Green"
+          title="Green [mod+5]"
         >
           ​🟢​​
         </button>
@@ -121,7 +152,7 @@ const markmapStore = useMarkmap()
           data-color="#2656bf"
           id="pen-color-blue"
           aria-label="Blue"
-          title="Blue"
+          title="Blue [mod+6]"
         >
           ​​🔵​​
         </button>
@@ -130,7 +161,7 @@ const markmapStore = useMarkmap()
           data-color="#c24aed"
           id="pen-color-purple"
           aria-label="Purple"
-          title="Purple"
+          title="Purple [mod+7]"
         >
           ​🟣​​
         </button>
@@ -139,7 +170,7 @@ const markmapStore = useMarkmap()
           data-color="#bf6b26"
           id="pen-color-brown"
           aria-label="Brown"
-          title="Brown"
+          title="Brown [mod+8]"
         >
           ​​🟤​
         </button>
@@ -148,7 +179,7 @@ const markmapStore = useMarkmap()
           class="pen-button"
           id="bg-white"
           aria-label="Toggle White Background"
-          title="Toggle White Background"
+          title="Toggle White Background [mod+9]"
         >
           ⬜
         </button>
@@ -156,7 +187,7 @@ const markmapStore = useMarkmap()
           class="pen-button"
           id="bg-black"
           aria-label="Toggle Black Background"
-          title="Toggle Black Background"
+          title="Toggle Black Background [mod+0]"
         >
           ⬛️
         </button>
@@ -166,17 +197,17 @@ const markmapStore = useMarkmap()
           type="range"
           min="1"
           max="10"
-          value="1"
+          value="5"
           step="0.5"
           name="Size"
-          title="Size"
+          title="Size [alt+(+/-)]"
         />
         <div class="mt-1 mx-2 opacity-25">/</div>
         <button
           class="pen-button active"
           id="l-solid"
           aria-label="Solid"
-          title="Solid"
+          title="Solid [ctrl+alt+1]"
         >
           —
         </button>
@@ -184,7 +215,7 @@ const markmapStore = useMarkmap()
           class="pen-button"
           id="l-dashed"
           aria-label="Dashed"
-          title="Dashed"
+          title="Dashed [ctrl+alt+2]"
         >
           ┅
         </button>
@@ -192,12 +223,12 @@ const markmapStore = useMarkmap()
           class="pen-button"
           id="l-dotted"
           aria-label="Dotted"
-          title="Dotted"
+          title="Dotted [ctrl+alt+3]"
         >
           ⋯
         </button>
         <div class="mt-1 mx-2 opacity-25">/</div>
-        <button class="pen-button" id="close" title="Close">❌</button>
+        <button class="pen-button" id="close" title="Close [mod+p]">❌</button>
       </div>
     </div>
 
@@ -215,7 +246,8 @@ const markmapStore = useMarkmap()
   overflow: hidden;
   margin: 0;
   padding: 0;
-  cursor: url('data:image/svg+xml,%3Csvg%20t%3D%271664251073231%27%20class%3D%27icon%27%20viewBox%3D%270%200%201024%201024%27%20version%3D%271.1%27%20xmlns%3D%27http%3A%2F%2Fwww.w3.org%2F2000%2Fsvg%27%20p-id%3D%279380%27%20width%3D%2716%27%20height%3D%2716%27%3E%3Cpath%20d%3D%27M508.202667%20270.08l245.76%20245.76-337.066667%20337.493333a107.562667%20107.562667%200%200%201-57.6%2029.866667l-202.24%2033.706667a41.386667%2041.386667%200%200%201-7.253333%200.426666%2043.392%2043.392%200%200%201-30.293334-12.8%2043.861333%2043.861333%200%200%201-12.373333-37.546666l33.706667-202.24A107.562667%20107.562667%200%200%201%20170.666667%20607.146667zM866.133333%20157.866667a173.056%20173.056%200%200%200-245.76%200l-51.626666%2051.626666%20245.76%20245.76%2051.626666-51.626666a173.056%20173.056%200%200%200%200-245.76z%27%20fill%3D%27%2341416E%27%20p-id%3D%279381%27%3E%3C%2Fpath%3E%3C%2Fsvg%3E')
+  cursor:
+    url('data:image/svg+xml,%3Csvg%20t%3D%271664251073231%27%20class%3D%27icon%27%20viewBox%3D%270%200%201024%201024%27%20version%3D%271.1%27%20xmlns%3D%27http%3A%2F%2Fwww.w3.org%2F2000%2Fsvg%27%20p-id%3D%279380%27%20width%3D%2716%27%20height%3D%2716%27%3E%3Cpath%20d%3D%27M508.202667%20270.08l245.76%20245.76-337.066667%20337.493333a107.562667%20107.562667%200%200%201-57.6%2029.866667l-202.24%2033.706667a41.386667%2041.386667%200%200%201-7.253333%200.426666%2043.392%2043.392%200%200%201-30.293334-12.8%2043.861333%2043.861333%200%200%201-12.373333-37.546666l33.706667-202.24A107.562667%20107.562667%200%200%201%20170.666667%20607.146667zM866.133333%20157.866667a173.056%20173.056%200%200%200-245.76%200l-51.626666%2051.626666%20245.76%20245.76%2051.626666-51.626666a173.056%20173.056%200%200%200%200-245.76z%27%20fill%3D%27%2341416E%27%20p-id%3D%279381%27%3E%3C%2Fpath%3E%3C%2Fsvg%3E')
       0 12,
     pointer;
 }